fix: fixes made after running PHPStan with level 5

+++ src/routes-management.php
@@ -3,36 +3,42 @@
 declare(strict_types=1);
 
 use BlogPostsHandling\Api\Controller\HomeController;
-use BlogPostsHandling\Api\Controller\{AddPostController,GetPostController,UpdatePostController,DeletePostController};
+use BlogPostsHandling\Api\Controller\{AddPostController,GetPostController,
+    UpdatePostController,DeletePostController};
 use BlogPostsHandling\Api\Controller\GetPostsBySlugController;
-use BlogPostsHandling\Api\Controller\{AddCategoryController,GetCategoryController,UpdateCategoryController,DeleteCategoryController};
+use BlogPostsHandling\Api\Controller\{AddCategoryController,GetCategoryController,
+    UpdateCategoryController,DeleteCategoryController};
 use BlogPostsHandling\Api\Controller\{ListCategoriesController,AddCategoryToAPostController};
 use BlogPostsHandling\Api\Controller\OpenApiDocController;
 
 /* Application routes  */
-$app->get('/v1', HomeController::class);
+if (isset($app)) /* added upon suggestion by PHPStan*/
+{
 
-$app->post('/v1/new/category', AddCategoryController::class);
+    $app->get('/v1', HomeController::class);
 
-$app->get('/v1/read/category/{id}', GetCategoryController::class);
+    $app->post('/v1/new/category', AddCategoryController::class);
 
-$app->put('/v1/update/category/{id}', UpdateCategoryController::class);
+    $app->get('/v1/read/category/{id}', GetCategoryController::class);
 
-$app->get('/v1/list/categories', ListCategoriesController::class);
+    $app->put('/v1/update/category/{id}', UpdateCategoryController::class);
 
-$app->delete('/v1/delete/category/{id}', DeleteCategoryController::class);
+    $app->get('/v1/list/categories', ListCategoriesController::class);
 
-$app->post('/v1/new/post', AddPostController::class);
+    $app->delete('/v1/delete/category/{id}', DeleteCategoryController::class);
 
-$app->get('/v1/read/post/{id}', GetPostController::class);
+    $app->post('/v1/new/post', AddPostController::class);
 
-$app->put('/v1/update/post/{id}', UpdatePostController::class);
+    $app->get('/v1/read/post/{id}', GetPostController::class);
 
-$app->delete('/v1/delete/post/{id}', DeletePostController::class);
+    $app->put('/v1/update/post/{id}', UpdatePostController::class);
 
-$app->get('/v1/posts/slug/{slug}', GetPostsBySlugController::class);
+    $app->delete('/v1/delete/post/{id}', DeletePostController::class);
 
-$app->post('/v1/post/{pid}/addto/{cid}', AddCategoryToAPostController::class);
+    $app->get('/v1/posts/slug/{slug}', GetPostsBySlugController::class);
 
-$app->get('/v1/apidocs', OpenApiDocController::class);
+    $app->post('/v1/post/{pid}/addto/{cid}', AddCategoryToAPostController::class);
+
+    $app->get('/v1/apidocs', OpenApiDocController::class);
+}
 /* Application routes ends */
